refactor : 사진첨부 부분 수정
